% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFMIndex.R
\name{computeFMIndex}
\alias{computeFMIndex}
\title{Writes the FM index}
\usage{
computeFMIndex(
  fastafile,
  usrpath,
  tallywidth = 1,
  includeTerminationCharacter = TRUE
)
}
\arguments{
\item{fastafile}{The .fasta file containing one single DNA sequence, files containing multiple sequences are not allowed}

\item{usrpath}{The path where the user wants to save the files.}

\item{tallywidth}{How you want to compress the tally table. Default value=1}

\item{includeTerminationCharacter}{default value: true}
}
\description{
This function writes the FM index of the DNA sequence contained in the .fasta file given in input
}
\examples{
computeFMIndex(system.file("extdata/seq5.fasta",package="FMIndex"),"test01.txt")
computeFMIndex(system.file("extdata/seq5.fasta",package="FMIndex"),"test02.txt",tallywidth=1)
}
