% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TallyTable.R
\name{getTallyTableLine}
\alias{getTallyTableLine}
\title{Create the i-th row of the Tally Table}
\usage{
getTallyTableLine(index, l.col, alphabet)
}
\arguments{
\item{index}{An integer representing the index of the i-th row you want
to build}

\item{l.col}{A DNAString object containing the L column of the FM index}

\item{alphabet}{A character vector containing the alphabet to consider in
counting the elements}
}
\value{
the i-th row of the Tally Table
}
\description{
This function creates the i-th row of the Tally Table
}
